requirements/runtime.txt
transformers
sentencepiece
datasets
numpy
tqdm
einops
psutil
packaging
pre-commit
ninja
gputil
pytest
boto3
botocore
torch-scatter
pyecharts
py-libnuma
pynvml
tensorboard
-f https://data.pyg.org/whl/torch-2.1.0+cu118.html
